added tests, fixed bug in query
I've added tests for yahoo.yql and yahoo.query. Unfortunately, these are
not comprehensive, but I'm sure I'll get there. Also, there was a bug in
yahoo.query/query in which keywords could be out of order. The problem
was due to 'conj' acting like 'cons' when presented with a list instead
of a vector.